backend/requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
Authlib==1.6.4
autopep8==2.3.2
bandit==1.8.6
beautifulsoup4==4.13.5
billiard==4.2.1
black==25.9.0
blinker==1.9.0
celery==5.5.3
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
configparser==7.2.0
contourpy==1.3.3
coverage==7.10.7
cryptography==46.0.2
cycler==0.12.1
Deprecated==1.2.18
distlib==0.3.9
dotenv==0.9.9
dparse==0.6.4
extensions==0.4
Faker==37.4.2
filelock==3.18.0
flake8==7.3.0
flasgger==0.9.7.1
Flask==2.3.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
fonttools==4.60.0
fsspec==2025.9.0
future==1.0.0
gensim==4.3.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kombu==5.5.4
limits==5.4.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.1.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyotp==2.9.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-whois==0.8.0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==13.9.4
rpds-py==0.26.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
safety==3.6.2
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.35.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stevedore==5.5.0
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
Tree==0.2.4
trio==0.30.0
trio-websocket==0.12.2
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
